Regroup the files to match closer to addressbook style
