+++ src/main/java/duke/Main.java
@@ -1,0 +1,25 @@
+package duke;
+
+import javafx.application.Application;
+import javafx.fxml.FXMLLoader;
+import javafx.scene.Scene;
+import javafx.scene.layout.AnchorPane;
+import javafx.stage.Stage;
+
+import java.io.IOException;
+
+public class Main extends Application {
+    @Override
+    public void start(Stage primaryStage) {
+        try{
+            FXMLLoader fxmlLoader = new FXMLLoader();
+            fxmlLoader.setLocation(getClass().getResource("/view/interface.fxml"));
+            AnchorPane pane = fxmlLoader.load();
+            Scene scene = new Scene(pane);
+            primaryStage.setScene(scene);
+            primaryStage.show();
+        } catch (IOException e){
+            e.printStackTrace();
+        }
+    }
+}
